-- flock main --

+++ flockinventory.py
@@ -23,8 +23,11 @@ def run():
     owls = loadinventory()
     for owl in owls: 
         flogger("checks for owl name -> %s, owl ip -> %s" % (owl["name"], owl["ip"]))
-        flockmonitor.check_owl_alive(owl)
-        flockmonitor.get_status_cpu(owl)
-        flockmonitor.remove_file(owl, "pcap.file")
+        alive, ssh = flockmonitor.check_owl_alive(owl)
+        if alive:
+            flockmonitor.get_status_sniffer(owl, ssh)
+            flockmonitor.get_status_cpu(owl)
+            flockmonitor.remove_file(owl, "pcap.file")
+            ssh.close()
